LayoutTests/http/tests/websocket/tests/hybi/send-empty-expected.txt

WebSocket: Calling send() without arguments should raise SyntaxError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.send() threw exception SyntaxError: Not enough arguments.
PASS closeEvent.wasClean is true
PASS receivedMessages.length is 1
PASS receivedMessages[0] is "Goodbye"
PASS successfullyParsed is true

TEST COMPLETE
